+++ packages/dialect-wasqlite-worker/src/index.ts
@@ -9,11 +9,13 @@ export * from './worker/utils'
 
 export class WaSqliteWorkerDialect extends GenericSqliteWorkerDialect<globalThis.Worker, InitData> {
   constructor(config: WaSqliteWorkerDialectConfig) {
-    const { onCreateConnection, worker, fileName, preferOPFS, url } = config
+    const { onCreateConnection, worker, fileName, preferOPFS, url, message } = config
     super(
       async () => {
         const supportModule = isModuleWorkerSupport()
         const useOPFS = preferOPFS ? await isOpfsSupported() : false
+        const m = mitt()
+        await message?.(m)
         return {
           data: {
             fileName,
@@ -27,7 +29,7 @@ export class WaSqliteWorkerDialect extends GenericSqliteWorkerDialect<globalThis
             : supportModule
               ? new Worker(new URL('worker.mjs', import.meta.url), { type: 'module' })
               : new Worker(new URL('worker.js', import.meta.url)),
-          mitt: mitt(),
+          mitt: m,
           handle: handleWebWorker,
         }
       },
